requirements.txt
audioread==3.0.1
azure-cognitiveservices-speech==1.20.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
exceptiongroup==1.2.2
fonttools==4.54.1
future==1.0.0
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
joblib==1.4.2
kiwisolver==1.4.7
librosa==0.9.2
llvmlite==0.31.0
matplotlib==3.7.5
numba==0.48.0
numpy==1.23.0
packaging==24.1
pandas==2.0.3
PeakUtils==1.3.5
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
pycparser==2.22
pydub==0.25.1
pyloudnorm==0.1.0
pyparsing==3.1.4
pysptk==1.0.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
resampy==0.3.1
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
surfboard==0.2.0
threadpoolctl==3.5.0
tomli==2.0.2
tqdm==4.66.5
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
